# Rewrite `Quantity` multiplication and division helpers to be faster