@mhvk 2025-05-14
I think the speed-up is most impressive for less likely cases, so I don't think we need to advertise it. But it is still very nice, of course!