@eerovaher 2025-05-14
Responding to https://github.com/astropy/astropy/pull/18146#pullrequestreview-2838848365 by @mhvk

> I would not have thought it would make this much of a difference, probably partially because the `_d(unit)` looks like nothing but does cost time.

The speedup does not come from avoiding `_d()`, the speedup comes from avoiding unit arithmetic if it is not needed.

> It does make me wonder whether part of the reason this was slow is that multiplying/dividing a unit by dimensionless is currently slow...

Technically it would be very simple to speed up the code noticeably by turning the `is_unity()` methods the different unit classes have into cached properties, but turning a method into a property would break public API. Caching the method is not a good idea because the cache will maintain references to unit instances, which prevents them from being garbage-collected and can therefore cause memory leaks.

Responding to https://github.com/astropy/astropy/pull/18146#issuecomment-2880020993 by @pllim 

> Are you sure this doesn't need a perf change log? Might be nice to advertise.

In practice it is common enough that the performance of some piece of code is affected by more than one pull request, which means that the change log entries for individual pull requests are either misleading or too vague to be of any use to anyone. At this point I'm convinced that announcing performance improvements in the change log is a bad idea. If there is something worth announcing, it should be done in a What's New entry instead.