@nstarman 2025-05-14
> avoiding unit arithmetic if it is not needed

In looking at the code, do you think there's opportunities to significantly speed up unit arithmetic? e.g. fast-paths, etc so that we don't have to optimize higher-up functions to avoid unit arithmetic? As in, can we push these optimizations a level down?