@mhvk 2025-05-14
On further thought (and influenced by @eerovaher's comments), the location here seems pretty spot on - usually, `None` does not necessarily mean unitless.