@nstarman 2025-05-14
> But the fast paths are not as fast as they could be because is_unity() is not cached, and I already commented on that in my previous post.

We could make a private cached property which is_unity() just returns, then add a deprecation warning to the method, then change it.