@mhvk 2025-05-14
> > But the fast paths are not as fast as they could be because is_unity() is not cached, and I already commented on that in my previous post.
> 
> We could make a private cached property which is_unity() just returns, then add a deprecation warning to the method, then change it.

Personally, I'd rather get rid of `is_unity()` altogether - it seems the wrong approach! Just like some units are irreducible, there can be one unit that is dimensionless unscaled. So I think it would be substantially cleaner and simpler to ensure that `dimensionless_unscaled` is a special unit or at least a singleton (e.g., by defining a `CompositeUnit.__new__` which can return it as needed).